homework4

+++ class3/homework_3.py
@@ -19,8 +19,9 @@ EX:
 請輸入身高:1.7
 請輸入體重:50
 你的BMI為17.301038062283737
+
 """
-h = int(input("請輸入你的身高:"))
-w = int(input("請輸入你的體重:"))
+h = float(input("請輸入你的身高:"))
+w = float(input("請輸入你的體重:"))
 bmi = w / h**2
 print(bmi)
